OSIS-4728 moved UniversityDomainsLookup class into the new Training form file
